refactor(Project Structure): make resources feature
