+++ app/src/main/java/com/monstarlab/core/data/network/responses/ResourcesResponse.kt
@@ -1,6 +1,6 @@
 package com.monstarlab.core.data.network.responses
 
-import com.monstarlab.core.data.network.dtos.ResourceDto
+import com.monstarlab.features.resources.data.ResourceDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
