+++ app/src/main/java/com/monstarlab/core/data/storage/ResourcePreferenceStore.kt
@@ -3,7 +3,7 @@ package com.monstarlab.core.data.storage
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.monstarlab.arch.data.SharedPreferenceDataStore
-import com.monstarlab.core.domain.model.Resource
+import com.monstarlab.features.resources.domain.Resource
 import javax.inject.Inject
 
 class ResourcePreferenceStore @Inject constructor(
