+++ app/src/main/java/com/monstarlab/features/resources/data/ResourceDto.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.core.data.network.dtos
+package com.monstarlab.features.resources.data
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
