+++ app/src/main/java/com/monstarlab/features/resources/data/ResourceMapping.kt
@@ -1,6 +1,7 @@
 package com.monstarlab.core.data.network.dtos
 
-import com.monstarlab.core.domain.model.Resource
+import com.monstarlab.features.resources.domain.Resource
+import com.monstarlab.features.resources.data.ResourceDto
 
 fun ResourceDto.toEntity(): Resource {
     return Resource(
