+++ app/src/main/java/com/monstarlab/features/resources/domain/ResourceRepository.kt
@@ -1,11 +1,10 @@
-package com.monstarlab.core.data.repositories
+package com.monstarlab.features.resources.domain
 
 import com.monstarlab.arch.data.Repository
 import com.monstarlab.arch.extensions.mapSuccess
 import com.monstarlab.core.data.network.Api
 import com.monstarlab.core.data.network.dtos.toEntity
 import com.monstarlab.core.data.storage.ResourcePreferenceStore
-import com.monstarlab.core.domain.model.Resource
 import javax.inject.Inject
 
 class ResourceRepository @Inject constructor(
