+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/resourceFragment"
-        android:name="com.monstarlab.features.resources.ResourceFragment"
+        android:name="com.monstarlab.features.resources.ui.ResourceFragment"
         tools:layout="@layout/fragment_resource"
         android:label="ResourceFragment" />
